+++ tests/unit/core/utils/test_url.js
@@ -1,8 +1,30 @@
 import url from 'url';
 
-import { addQueryParams } from 'core/utils/url';
+import { addQueryParams, removeUndefinedProps } from 'core/utils/url';
 
 describe(__filename, () => {
+  describe('removeUndefinedProps', () => {
+    it('removes undefined properties', () => {
+      expect(removeUndefinedProps({ thing: undefined })).toEqual({});
+    });
+
+    it('preserves falsy properties', () => {
+      expect(removeUndefinedProps({ thing: false })).toEqual({ thing: false });
+    });
+
+    it('preserves other properties', () => {
+      expect(removeUndefinedProps({ thing: 'thing' })).toEqual({
+        thing: 'thing',
+      });
+    });
+
+    it('does not modify the original object', () => {
+      const example = { thing: undefined };
+      removeUndefinedProps(example);
+      expect(example).toEqual({ thing: undefined });
+    });
+  });
+
   describe('addQueryParams', () => {
     it('adds a query param to a plain url', () => {
       const output = addQueryParams('http://whatever.com/', { foo: 'bar' });
